+++ src/main/java/hello/controllers/EarthquakeController.java
@@ -13,7 +13,7 @@ import java.util.HashMap;
 
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 import hello.geojson.FeatureCollection;
-import hello.EarthquakeQueryService;
+import hello.services.EarthquakeQueryService;
 import hello.EqSearch;
